refactor(merge): split merge.rs into a module and decompose reconcile_cluster (#945)
Behavior-preserving refactor of the 12.7k-line merge.rs, the #938
compaction-lane serialization bottleneck. No production behavior change —
compaction output and the #1037 purge tally are byte-for-byte identical.

- Convert `merge.rs` to a `merge/` module (`mod.rs`). External
  `write_engine::merge::{MergeEntry, MergeStep, RowData, KWayMerger, ...}`
  paths keep resolving via re-exports, so no call sites change.
- Extract the data model (`MergeEntry`, `RowData`, `CellData`,
  `ComplexDeletion`, `MergeStep`, `MergeStats` + builders) into
  `merge/model.rs`, re-exported. Pure relocation; every item keeps its exact
  `#[cfg(feature = "write-support")]` gate. `RangeCut` stays in `mod.rs` (it
  belongs with the range-coalescing machinery, not the public model).
- Decompose the ~470-line `reconcile_cluster_with_overlap_counted` into
  named, separately-testable steps on a `ReconcileState` accumulator in
  `merge/reconcile.rs`: fold_row_deletions, resolve_cell_winners,
  apply_complex_deletions, shadow_by_row_deletion, filter_dropped_columns,
  purge_gc_grace (threads the #1037 PurgeCounts tally), build. The step
  bodies are lifted verbatim; the parity-load-bearing order (Step 2b before
  3/3c; `had_data_before` captured pre-purge, consulted post-purge) is
  preserved. The orchestrator method stays
  `KWayMerger::reconcile_cluster_with_overlap_counted`, so the 123 in-file
  tests are untouched.

No visibility changes: the reconcile steps reach the private cell helpers and
`PurgeCounts` as a child module of `merge`.

Validated: cqlite-core lib tests (2284 pass), 123 merge tests preserved,
differential compaction + range-tombstone + integration parity tests pass,
clippy --all-features --all-targets clean, minimal (no-write-support) build
clean, fmt clean.

Per issue #945 Comment 4 (highest-ratio slice): defers the heap/adapter file
extractions and test relocation as follow-up polish.
