+++ cqlite-core/src/storage/write_engine/merge/mod.rs
@@ -70,374 +70,14 @@ use std::path::{Path, PathBuf};
 #[cfg(feature = "write-support")]
 use std::time::{Duration, Instant};
 
-/// Entry in the merge stream
-///
-/// Represents a single row from one of the input SSTables. This is the
-/// fundamental unit that flows through the merge heap.
+mod model;
 #[cfg(feature = "write-support")]
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct MergeEntry {
-    /// Which SSTable this came from (0 = newest)
-    pub run_index: usize,
-    /// Partition key with token
-    pub key: DecoratedKey,
-    /// Clustering key (None for tables without clustering)
-    pub clustering_key: Option<ClusteringKey>,
-    /// Timestamp in microseconds since Unix epoch
-    pub timestamp: i64,
-    /// Row data (live cells or tombstone)
-    pub row_data: RowData,
-    /// Complex (collection / UDT) deletion markers for the multi-cell columns
-    /// of this `(pk, ck)` (issue #886 substrate).
-    ///
-    /// Carried through the merge so the per-cell-path collection/UDT followup
-    /// (#844) and shadow-before-purge (#887) can preserve collection/UDT
-    /// deletion timestamps. `reconcile_cluster` unions and preserves these
-    /// across a cluster, but reconciliation does **not yet consult** them and
-    /// the writer does not yet apply them — defaults to empty, so output is
-    /// byte-unchanged. Population (per-element reader emit) lands in #899.
-    pub complex_deletions: Vec<ComplexDeletion>,
-    /// Range-deletion marker covering a span of clustering keys (issue #886
-    /// substrate).
-    ///
-    /// A first-class slot so range tombstones can flow through the merge stream
-    /// instead of being skipped by the parser; applying them to shadow covered
-    /// cells is the follow-up #846. Carried (and timestamp-max-preserved)
-    /// through `reconcile_cluster` but **not yet consulted** by reconciliation
-    /// or the writer, so output is byte-unchanged. `None` when this entry
-    /// carries no range deletion.
-    pub range_deletion: Option<RangeTombstone>,
-    /// Row-level deletion that COEXISTS with this entry's surviving live cells
-    /// (issue #932).
-    ///
-    /// `Some((deletion_time_micros, local_deletion_time_secs))` when a row
-    /// tombstone whose timestamp is OLDER than the surviving cells must be kept
-    /// alongside `RowData::Live`. `RowData::Live` has no row-deletion slot, so —
-    /// mirroring the carry-only `complex_deletions` / `range_deletion` fields —
-    /// the coexisting deletion rides here instead of forcing a `RowData` change
-    /// across ~100 `RowData::Live` sites. `reconcile_cluster` folds it into the
-    /// winning row deletion and re-attaches it to the emitted live entry;
-    /// `merge_entry_to_mutation` emits it as `Mutation::row_tombstone`. Without
-    /// this, a partial compaction that keeps newer cells would DROP the row
-    /// deletion and let older cells of other columns (in SSTables not part of the
-    /// compaction) resurrect. `None` for a row with no coexisting deletion.
-    pub row_deletion: Option<(i64, i32)>,
-    /// Partition-level deletion `(markedForDeleteAt µs, localDeletionTime s)`
-    /// carried by a synthetic partition-tombstone carrier entry (issue #1072).
-    ///
-    /// `Some(..)` ONLY on the carrier entry the reader emits for a partition
-    /// header that has a tombstone (empty `RowData::Live`, `clustering_key:
-    /// None`). The merge extracts the MAX `markedForDeleteAt` across sources,
-    /// applies it as the OUTERMOST per-cell `<=` shadow floor for the whole
-    /// partition, and re-emits the surviving tombstone via
-    /// `merge_entry_to_mutation` (→ `Mutation::partition_tombstone`). Without
-    /// this a newer partition tombstone in one SSTable failed to shadow older
-    /// live rows in another, resurrecting the deleted partition. `None` for every
-    /// non-carrier entry.
-    pub partition_deletion: Option<(i64, i32)>,
-}
+pub use model::{CellData, ComplexDeletion, MergeEntry, MergeStats, MergeStep, RowData};
 
-impl MergeEntry {
-    /// Create a new merge entry.
-    ///
-    /// The carry-only #886 substrate fields (`complex_deletions`,
-    /// `range_deletion`) default to empty/`None`; attach them with
-    /// [`with_complex_deletions`](Self::with_complex_deletions) /
-    /// [`with_range_deletion`](Self::with_range_deletion) once the reader emit
-    /// surfaces them (#899).
-    pub fn new(
-        run_index: usize,
-        key: DecoratedKey,
-        clustering_key: Option<ClusteringKey>,
-        timestamp: i64,
-        row_data: RowData,
-    ) -> Self {
-        Self {
-            run_index,
-            key,
-            clustering_key,
-            timestamp,
-            row_data,
-            complex_deletions: Vec::new(),
-            range_deletion: None,
-            row_deletion: None,
-            partition_deletion: None,
-        }
-    }
-
-    /// Attach complex-deletion markers (issue #886 substrate; carry-only).
-    #[must_use]
-    pub fn with_complex_deletions(mut self, complex_deletions: Vec<ComplexDeletion>) -> Self {
-        self.complex_deletions = complex_deletions;
-        self
-    }
-
-    /// Attach a coexisting row-level deletion (issue #932).
-    ///
-    /// `deletion_time` is `markedForDeleteAt` (microseconds); `ldt` is the
-    /// `localDeletionTime` (GC-clock seconds, `0` = not surfaced). Used when a
-    /// row's surviving cells are newer than a row tombstone that must still be
-    /// preserved so it keeps shadowing older cells in non-compacted SSTables.
-    #[must_use]
-    pub fn with_row_deletion(mut self, deletion_time: i64, ldt: i32) -> Self {
-        self.row_deletion = Some((deletion_time, ldt));
-        self
-    }
-
-    /// Attach a range-deletion marker (issue #886 substrate; carry-only).
-    #[must_use]
-    pub fn with_range_deletion(mut self, range_deletion: RangeTombstone) -> Self {
-        self.range_deletion = Some(range_deletion);
-        self
-    }
-
-    /// Attach a partition-level deletion (issue #1072).
-    ///
-    /// `deletion_time` is `markedForDeleteAt` (microseconds); `ldt` is the
-    /// `localDeletionTime` (GC-clock seconds). Marks this entry as the synthetic
-    /// partition-tombstone carrier so the merge applies the partition floor and
-    /// re-emits the surviving tombstone.
-    #[must_use]
-    pub fn with_partition_deletion(mut self, partition_deletion: (i64, i32)) -> Self {
-        self.partition_deletion = Some(partition_deletion);
-        self
-    }
-
-    /// True when this entry has NO writer-emittable content at all — an empty
-    /// `RowData::Live { cells: vec![] }` with no row tombstone and no carried
-    /// deletion metadata. Routing such an entry to the writer would create a
-    /// PHANTOM live empty (pure-PK) row at timestamp 0 (`DataWriter::merge_row_group`
-    /// treats a no-op mutation as a primary-key insert), so `merge` filters it.
-    ///
-    /// Epic #899 Phase C: a COMPLEX DELETION is consumed by the writer (a real
-    /// `CellOperation::ComplexDeletion` marker), so a complex-deletion-only entry
-    /// is NOT a no-op.
-    ///
-    /// Issue #933: a RANGE DELETION is now ALSO consumed by the compaction writer —
-    /// `merge_entry_to_mutation` threads it onto the mutation's `range_tombstones`,
-    /// which the writer interleaves as on-disk bound markers AND uses to shadow
-    /// covered same-partition rows. A range-deletion-only carrier is therefore NO
-    /// LONGER a no-op and MUST reach the writer (dropping it would resurrect the
-    /// covered rows it shadowed — roborev #959 High #2). In practice
-    /// `reconcile_cluster` never emits a truly-empty entry (one with empty cells,
-    /// no row tombstone, and no carried metadata reconciles to `None`), so this
-    /// guard is now effectively unreachable but kept as a defensive phantom-row
-    /// filter.
-    #[must_use]
-    pub fn is_metadata_only_no_op(&self) -> bool {
-        matches!(&self.row_data, RowData::Live { cells } if cells.is_empty())
-            && self.complex_deletions.is_empty()
-            && self.range_deletion.is_none()
-            && self.row_deletion.is_none()
-            && self.partition_deletion.is_none()
-    }
-
-    /// True when this entry is a synthetic partition-tombstone carrier (issue
-    /// #1072): an empty live row whose only payload is `partition_deletion`.
-    /// Produced by [`Self::build_merge_entry`] from a reader `PartitionDelete`.
-    /// Such an entry produces NO output row (it yields a
-    /// `Mutation::partition_tombstone`, not a clustering row), so the merge must
-    /// not count it toward output row counts — but it MUST reach the writer.
-    #[must_use]
-    pub fn is_partition_delete_carrier(&self) -> bool {
-        self.partition_deletion.is_some()
-    }
-}
-
-/// Ord implementation for min-heap routing ONLY (not LWW winner selection).
-///
-/// This orders entries so the heap yields them grouped by partition and
-/// clustering key. The actual equal-timestamp Delete-vs-Live winner is chosen
-/// in `merge_partition_rows` (timestamp → liveness → run_index), NOT here.
-///
-/// Order by:
-/// 1. Token (ascending)
-/// 2. Key bytes (ascending, for hash collisions)
-/// 3. Clustering key (ascending, schema-aware)
-/// 4. Run index (ascending) - stable routing tiebreak only
+/// Clustering-group reconciliation kernel, decomposed into named steps
+/// (issue #945). See [`reconcile::ReconcileState`].
 #[cfg(feature = "write-support")]
-impl Ord for MergeEntry {
-    fn cmp(&self, other: &Self) -> Ordering {
-        // Primary: by token
-        match self.key.token.cmp(&other.key.token) {
-            Ordering::Equal => {
-                // Secondary: by key bytes (hash collision resolution)
-                match self.key.key.cmp(&other.key.key) {
-                    Ordering::Equal => {
-                        // Tertiary: by clustering key
-                        match (&self.clustering_key, &other.clustering_key) {
-                            (None, None) => {
-                                // Quaternary: by run_index (lower = newer)
-                                self.run_index.cmp(&other.run_index)
-                            }
-                            (None, Some(_)) => Ordering::Less,
-                            (Some(_), None) => Ordering::Greater,
-                            (Some(a), Some(b)) => {
-                                // Use fallback Ord (not schema-aware at this level)
-                                // Schema-aware comparison happens during partition merge
-                                match a.cmp(b) {
-                                    Ordering::Equal => {
-                                        // Equal clustering keys: prefer lower run_index
-                                        self.run_index.cmp(&other.run_index)
-                                    }
-                                    other_ord => other_ord,
-                                }
-                            }
-                        }
-                    }
-                    other_ord => other_ord,
-                }
-            }
-            other_ord => other_ord,
-        }
-    }
-}
-
-#[cfg(feature = "write-support")]
-impl PartialOrd for MergeEntry {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-/// Row data: live cells or tombstone
-#[cfg(feature = "write-support")]
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum RowData {
-    /// Live row with cell data
-    Live {
-        /// Cell data for this row
-        cells: Vec<CellData>,
-    },
-    /// Row tombstone
-    Tombstone {
-        /// Deletion timestamp (microseconds)
-        deletion_time: i64,
-        /// Local deletion time (seconds since epoch)
-        local_deletion_time: i32,
-    },
-}
-
-/// Cell data with timestamp, optional TTL, and (for complex columns) cell path.
-///
-/// ## Per-cell merge metadata (issue #886 — byte-parity foundation)
-///
-/// To reconcile per-cell and per-element data byte-faithfully (Cassandra
-/// `Cells#reconcile`), the merge entry must carry more than a single row-level
-/// timestamp. The fields below thread that richer state from the reader toward
-/// the followup behaviors (#844 per-cell-path collection/UDT merge, #848
-/// tombstone-vs-expiring TTL tie-break). They are **carried but not yet acted
-/// on** by reconciliation — this struct change is plumbing only and must not
-/// alter output bytes.
-///
-/// Where the reader does not yet surface a value the field is left `None`; the
-/// dependent issues fill it in once the reader is extended.
-#[cfg(feature = "write-support")]
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct CellData {
-    /// Column name
-    pub column: String,
-    /// Column value
-    pub value: Value,
-    /// Cell timestamp (microseconds)
-    pub timestamp: i64,
-    /// TTL in seconds (None = no expiration)
-    pub ttl: Option<u32>,
-    /// Cell path for a complex (collection / non-frozen UDT) element — the
-    /// serialized element key/index that distinguishes one element of a
-    /// multi-cell column from another (issue #886 substrate).
-    ///
-    /// **Carry-only.** This field is threaded through the merge entry so that
-    /// per-element reconciliation can become byte-faithful, but nothing
-    /// populates or consumes it yet: the reader still collapses collections to
-    /// a single whole-column [`CellData`] and the writer does not read this
-    /// field. Population (per-element reader emit) and consumption (per-path
-    /// merge #844) land in the follow-up #899. `None` for simple cells.
-    pub cell_path: Option<Vec<u8>>,
-    /// Local deletion time in **seconds** since the Unix epoch for this cell
-    /// (the on-disk `localDeletionTime`), used by gc_grace purging and
-    /// expiring-cell tie-breaks (issue #886 substrate).
-    ///
-    /// For an expiring (TTL) cell this is the cell's expiration instant; for a
-    /// cell tombstone it is when the delete was applied.
-    ///
-    /// Populated for complex (collection / UDT) elements on the compaction read
-    /// path (epic #899, Phase C); `None` for simple cells whose LDT the reader
-    /// does not surface, and for live simple cells.
-    pub local_deletion_time: Option<i32>,
-    /// True when this `CellData` represents a single ELEMENT of a non-frozen
-    /// complex column (a list/set member or a map entry), as opposed to a simple
-    /// single-cell column (epic #899, Phase C).
-    ///
-    /// When `true`, [`cell_path`](Self::cell_path) is the element's authoritative
-    /// on-disk path and the merge→mutation step emits a
-    /// [`CellOperation::WriteComplexElement`] (preserving per-element
-    /// ts/ttl/ldt/path) rather than a whole-column `Write`. `false` for every
-    /// simple cell (whole-column collapse no longer happens on the production
-    /// path).
-    ///
-    /// [`CellOperation::WriteComplexElement`]: crate::storage::write_engine::mutation::CellOperation::WriteComplexElement
-    pub is_complex_element: bool,
-    /// Authoritative IS_DELETED (0x01) flag for a complex element (epic #899,
-    /// Phase C). Carried verbatim from [`ComplexElement::is_deleted`] — NOT
-    /// re-derived from value/ttl/ldt shape, so an expiring SET member (empty
-    /// value, ttl + ldt set) is correctly NOT treated as a tombstone
-    /// (no-heuristics mandate). Always `false` for simple cells (a simple cell
-    /// tombstone rides in [`value`](Self::value) as `Value::Tombstone`).
-    ///
-    /// [`ComplexElement::is_deleted`]: crate::storage::sstable::reader::compaction_row::ComplexElement::is_deleted
-    pub is_deleted: bool,
-    /// On-disk HAS_EMPTY_VALUE (0x04) flag for a complex element (epic #899,
-    /// Phase C). `true` for a SET member (whose identity lives in the cell_path)
-    /// and any genuinely empty-value element, so the writer reproduces the same
-    /// on-disk emptiness rather than re-deriving it from the decoded value.
-    /// Always `false` for simple cells.
-    pub has_empty_value: bool,
-}
-
-#[cfg(feature = "write-support")]
-impl CellData {
-    /// Construct a simple live cell with no TTL, local-deletion-time, or cell
-    /// path. The richer fields default to `None`; populate them explicitly when
-    /// the reader supplies them (issues #844 / #848).
-    pub fn new(column: String, value: Value, timestamp: i64) -> Self {
-        Self {
-            column,
-            value,
-            timestamp,
-            ttl: None,
-            cell_path: None,
-            local_deletion_time: None,
-            is_complex_element: false,
-            is_deleted: false,
-            has_empty_value: false,
-        }
-    }
-}
-
-/// Complex (collection / non-frozen UDT) deletion marker for one column
-/// (issue #886 substrate).
-///
-/// Cassandra writes a complex-deletion marker ahead of a multi-cell column's
-/// elements to delete every element written at or before `marked_for_delete_at`.
-/// A merged complex deletion is dropped unless it **strictly supersedes** the
-/// active one (Cassandra commit `bd244649`). CQLite currently reduces this to a
-/// boolean and discards the timestamps; this first-class entity preserves them
-/// so per-path merge (#844) and shadow-before-purge (#887) can be byte-faithful.
-///
-/// **Carry-only.** Carried on [`MergeEntry`] (unioned through
-/// `reconcile_cluster`) but not yet populated by the reader or applied during
-/// the merge — that is the follow-up #899/#887.
-#[cfg(feature = "write-support")]
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ComplexDeletion {
-    /// Name of the complex column this deletion covers.
-    pub column: String,
-    /// Deletion timestamp (`markedForDeleteAt`) in microseconds since the epoch.
-    pub marked_for_delete_at: i64,
-    /// Local deletion time in seconds since the epoch.
-    pub local_deletion_time: i32,
-}
+mod reconcile;
 
 /// A cut position on the clustering axis, used to coalesce range tombstones
 /// into a NON-OVERLAPPING canonical sequence (issue #933 / roborev #959 High
@@ -474,37 +114,6 @@ enum RangeCut {
     },
     /// After all clustering keys (end of partition).
     Top,
-}
-
-/// Result of a merge step (incremental merge)
-#[cfg(feature = "write-support")]
-#[derive(Debug)]
-pub enum MergeStep {
-    /// Merged partition with all its rows
-    Partition {
-        /// Partition key
-        key: DecoratedKey,
-        /// All rows in this partition (already merged)
-        rows: Vec<MergeEntry>,
-    },
-    /// Merge is complete
-    Complete,
-}
-
-/// Statistics collected during merge
-#[cfg(feature = "write-support")]
-#[derive(Debug, Clone)]
-pub struct MergeStats {
-    /// Number of input files
-    pub input_files: usize,
-    /// Number of output partitions
-    pub output_partitions: u64,
-    /// Number of output rows
-    pub output_rows: u64,
-    /// Bytes written to output
-    pub bytes_written: u64,
-    /// Elapsed time
-    pub elapsed: Duration,
 }
 
 /// Buffered reader for a single SSTable run
@@ -3534,473 +3143,31 @@ impl KWayMerger {
         // (issue #1037). Never read here; only incremented.
         purges: &mut PurgeCounts,
     ) -> Option<MergeEntry> {
-        use std::collections::HashMap;
-
-        // Carry-through key fields: every entry in this group shares the same
-        // partition key and clustering key. Use the lowest run_index seen (newest
-        // file) so downstream ordering is stable.
-        let mut key = None;
-        let mut run_index = usize::MAX;
-
-        // Step 1: effective row deletion — max deletion_time across row tombstones.
-        // Track the source `local_deletion_time` (GC-clock seconds) PAIRED with the
-        // winning (max) deletion_time so the rebuilt tombstone preserves its
-        // wall-clock LDT instead of re-deriving it from the deletion timestamp
-        // (#873 — gc_grace semantics + unsigned LDT-delta underflow guard).
-        let mut row_del: Option<i64> = None;
-        let mut row_del_ldt: i32 = 0;
-
-        // Step 2: per-cell reconcile keyed by `(column, cell_path)` so each
-        // element of a multi-cell column reconciles independently (epic #899).
-        // Simple cells have `cell_path == None`, so their key is `(column,
-        // None)` and they behave exactly as before. Preserve first-seen key order
-        // for deterministic output while resolving winners in a side map.
-        type CellKey = (String, Option<Vec<u8>>);
-        let mut order: Vec<CellKey> = Vec::new();
-        let mut winners: HashMap<CellKey, CellData> = HashMap::new();
-
-        // Carried complex deletions (epic #899 substrate). Accumulate the inputs'
-        // markers as a first-seen, de-duplicated union here; the strict-supersede
-        // reduction (one max-mfda deletion per column NAME) and shadow-before-purge
-        // (dropping covered elements) run in Step 2b AFTER per-cell winner resolution
-        // and BEFORE the row-tombstone / dropped-column filters (issue #887, parity
-        // Cassandra `bd244649` + `f66fa14f`).
-        let mut complex_deletions: Vec<ComplexDeletion> = Vec::new();
-        let mut range_deletion: Option<RangeTombstone> = None;
-
-        for entry in &cluster_rows {
-            if key.is_none() {
-                key = Some(entry.key.clone());
-            }
-            run_index = run_index.min(entry.run_index);
-
-            for cd in &entry.complex_deletions {
-                if !complex_deletions.contains(cd) {
-                    complex_deletions.push(cd.clone());
-                }
-            }
-            if let Some(rd) = &entry.range_deletion {
-                let replace = match &range_deletion {
-                    None => true,
-                    Some(current) => rd.deletion_time > current.deletion_time,
-                };
-                if replace {
-                    range_deletion = Some(rd.clone());
-                }
-            }
-
-            // Issue #932: a coexisting row deletion carried on a LIVE entry (a
-            // read-back row that held BOTH a row tombstone and surviving newer
-            // cells) contributes to the winning row deletion exactly as a
-            // standalone `RowData::Tombstone` does — capture its paired LDT when
-            // it sets the new max so the rebuilt deletion preserves its wall-clock
-            // `localDeletionTime` (#873).
-            if let Some((del_ts, del_ldt)) = entry.row_deletion {
-                if row_del.is_none_or(|d| del_ts > d) {
-                    row_del = Some(del_ts);
-                    row_del_ldt = del_ldt;
-                }
-            }
-
-            match &entry.row_data {
-                RowData::Tombstone {
-                    deletion_time,
-                    local_deletion_time,
-                } => {
-                    // When this tombstone's deletion_time becomes (or sets) the new
-                    // max, capture its paired LDT too so the winning tombstone's
-                    // source `localDeletionTime` survives reconciliation (#873).
-                    if row_del.is_none_or(|d| *deletion_time > d) {
-                        row_del = Some(*deletion_time);
-                        row_del_ldt = *local_deletion_time;
-                    }
-                }
-                RowData::Live { cells } => {
-                    for cell in cells {
-                        let cell_key: CellKey = (cell.column.clone(), cell.cell_path.clone());
-                        match winners.get(&cell_key) {
-                            None => {
-                                order.push(cell_key.clone());
-                                winners.insert(cell_key, cell.clone());
-                            }
-                            Some(existing) => {
-                                // Higher timestamp wins. At EQUAL timestamp a cell
-                                // DELETION (tombstone) beats a LIVE or EXPIRING
-                                // (TTL) cell, decided BEFORE any localDeletionTime
-                                // compare (parity Cassandra `a62c749`,
-                                // `Cells#reconcile`; issue #848 / #498). At equal ts
-                                // + equal deletion-status, keep the first-seen
-                                // (newer file) winner. See `cell_reconcile_replace`.
-                                if Self::cell_reconcile_replace(cell, existing) {
-                                    winners.insert(cell_key, cell.clone());
-                                }
-                            }
-                        }
-                    }
-                }
-            }
+        // Decomposed into named, parity-load-bearing steps in `reconcile.rs`
+        // (issue #945). The step ORDER is critical: Step 2b before Steps 3/3c so
+        // a surviving complex deletion cannot resurrect a covered element on a
+        // later purge (`f66fa14f`); `had_data_before` is captured pre-purge and
+        // consulted post-purge (#921 finding 3). Behavior and the #1037 purge
+        // tally are byte-identical.
+        let mut state = reconcile::ReconcileState::new(clustering_key);
+        // Step 1: fold per-entry row/complex/range deletion metadata.
+        state.fold_row_deletions(&cluster_rows);
+        // Step 2: per-(column, cell_path) last-write-wins winner resolution.
+        state.resolve_cell_winners(&cluster_rows);
+        // Empty group => nothing to emit (original `let key = key?`).
+        if !state.has_key() {
+            return None;
         }
-
-        let key = key?; // empty group => nothing to emit
-
-        // Step 2b: COMPLEX-DELETION reconcile — strict-supersede + shadow-before-purge
-        // (issue #887). This runs AFTER per-cell winner resolution and BEFORE the
-        // row-tombstone / dropped-column filters, so an element shadowed by a
-        // surviving complex deletion cannot be resurrected by a later purge of the
-        // deletion marker.
-        //
-        // 1. STRICT SUPERSEDE (parity Cassandra `bd244649`): per complex column the
-        //    ACTIVE deletion is the one with the greatest `marked_for_delete_at`. A
-        //    merged deletion supersedes the active one only when its mfda is STRICTLY
-        //    GREATER — EQUAL timestamps do NOT supersede. Reduce the carried
-        //    first-seen union to ONE deletion per column NAME (the strict max),
-        //    preserving first-seen column order for deterministic output. Matched BY
-        //    NAME (`ComplexDeletion.column`), never by header identity (consistent
-        //    with #888's match-by-name substrate).
-        //
-        // 2. SHADOW BEFORE PURGE (parity Cassandra `f66fa14f`): for the surviving
-        //    deletion on a column, drop every per-element winner of THAT column
-        //    (matched by NAME, only complex ELEMENTS carrying a `cell_path`) whose own
-        //    timestamp is `<= marked_for_delete_at`. The boundary mirrors the
-        //    cell-vs-deletion rule (#498): an element with ts STRICTLY GREATER than
-        //    mfda survives; `<=` is shadowed. A simple cell sharing the name is never
-        //    collapsed by a complex marker.
-        if !complex_deletions.is_empty() {
-            // Strict-supersede: collapse to the max-mfda deletion per column name.
-            let mut active: HashMap<String, ComplexDeletion> = HashMap::new();
-            let mut active_order: Vec<String> = Vec::new();
-            for cd in complex_deletions.drain(..) {
-                match active.get_mut(&cd.column) {
-                    None => {
-                        active_order.push(cd.column.clone());
-                        active.insert(cd.column.clone(), cd);
-                    }
-                    // STRICTLY GREATER supersedes; equal/lesser does NOT (bd244649).
-                    Some(existing) if cd.marked_for_delete_at > existing.marked_for_delete_at => {
-                        *existing = cd;
-                    }
-                    Some(_) => {}
-                }
-            }
-
-            // Shadow-before-purge: drop covered per-element winners of each column
-            // whose ts <= mfda (matched by NAME), so purging the marker later cannot
-            // resurrect them (f66fa14f). Elements strictly newer than mfda survive.
-            for column in &active_order {
-                if let Some(cd) = active.get(column) {
-                    let mfda = cd.marked_for_delete_at;
-                    order.retain(|cell_key| {
-                        let (cell_column, cell_path) = cell_key;
-                        // Only complex elements (those with a cell_path) of THIS
-                        // column are candidates for shadowing.
-                        if cell_column != column || cell_path.is_none() {
-                            return true;
-                        }
-                        match winners.get(cell_key) {
-                            // ts > mfda survives; ts <= mfda is shadowed (purged).
-                            Some(cell) if cell.timestamp > mfda => true,
-                            Some(_) => {
-                                winners.remove(cell_key);
-                                false
-                            }
-                            None => false,
-                        }
-                    });
-                }
-            }
-
-            // Rebuild the carried union as the surviving per-column deletions in
-            // first-seen column order.
-            complex_deletions = active_order
-                .into_iter()
-                .filter_map(|column| active.remove(&column))
-                .collect();
-        }
-
-        // Step 3: apply row-tombstone shadowing per cell. A cell whose timestamp is
-        // <= row_del is shadowed (`<=` lets the tombstone win at equal ts, #498).
-        // Cells written strictly after row_del survive. This shadowing applies to
-        // cell tombstones too: a row tombstone at ts=T supersedes a cell tombstone at
-        // ts<=T (real Cassandra semantics). Note this is INTENTIONALLY stricter than
-        // the `reference_merge` model, whose range-tombstone path only suppresses
-        // live cells — `reconcile_cluster` is the authoritative behavior here.
-        //
-        // COMPLEX-DELETION shadowing (dropping elements covered by a collection
-        // `markedForDeleteAt`) has already been applied in Step 2b above, BEFORE this
-        // row-tombstone filter, so a surviving complex deletion cannot resurrect a
-        // covered element on a later purge (issue #887, parity `f66fa14f`).
-        //
-        // Step 3b: dropped-column filtering (Cassandra `cb34ad47`,
-        // `compaction.purge`). A column dropped at `drop_time` discards every cell
-        // whose `timestamp <= drop_time`; a cell written strictly after the drop
-        // (the column was re-added) survives. Scoped per column via the
-        // `dropped_columns` map (#904 plumbing, #847 filter). `compact_sstables`
-        // keeps any dropped column with survivors in the *writer* schema and
-        // strips only the fully-purged ones from the output serialization header
-        // (`TableSchema::for_compaction_output`).
-        //
-        // EXACT PER-CELL GRANULARITY (#922): `cell.timestamp` is the cell's OWN
-        // writetime. The production compaction path builds each simple cell from
-        // the reader's per-cell `CellWriteMetadata.write_timestamp_micros`
-        // (`build_compaction_row_data` → `SimpleCell.timestamp`, the #886/#899
-        // enrichment), so a row mixing a pre-drop dropped-column cell with a
-        // post-drop cell of ANOTHER column purges the dropped cell by its own
-        // writetime while the sibling survives by its — even though they share one
-        // on-disk row. (Cells with no per-cell metadata fall back to the row
-        // timestamp, matching Cassandra when every cell shares a writetime.) The
-        // legacy `value_to_row_data` row-timestamp path is `#[cfg(test)]`-only and
-        // is not on the production compaction stream. Verified end-to-end by
-        // `tests/issue_922_per_cell_dropped_purge.rs`.
-        //
-        // Apply row-tombstone shadowing first, then dropped-column filtering as a
-        // second stage so we can tell whether the dropped-column purge is what
-        // emptied the row of real data (phantom-row guard below).
-        let after_row_del: Vec<CellData> = order
-            .into_iter()
-            .filter_map(|cell_key| winners.remove(&cell_key))
-            .filter(|cell| match row_del {
-                Some(d) => cell.timestamp > d,
-                None => true,
-            })
-            .collect();
-
-        let mut surviving: Vec<CellData> = after_row_del
-            .iter()
-            .filter(|cell| match dropped_columns.get(&cell.column) {
-                Some(drop_time) => cell.timestamp > *drop_time,
-                None => true,
-            })
-            .cloned()
-            .collect();
-
-        // Phantom-row guard (#847 review): clustering-key columns are intentionally
-        // left in the cell list (see `extract_clustering_key`) for read-back. If a
-        // clustered row's only real (non-key) data is a dropped column (or, per
-        // #921 finding 3 below, a purgeable cell tombstone), the filters remove it
-        // but the clustering-key pseudo-cells remain — which would emit a phantom
-        // live row with a key but no data (the writer, whose schema excludes the
-        // dropped column, would serialize a key-only empty row). Suppress that:
-        // when the row HAD non-key data before the purges and has none after,
-        // treat it as data-less. A row that was always key-only (a genuine row
-        // marker) is preserved.
-        //
-        // We capture `had_data_before` here, but DEFER the
-        // `has_data_after` / `purged_to_empty` determination until AFTER the
-        // gc-grace tombstone purge (Step 3c) — see #921 finding 3 below. Computing
-        // it now (before Step 3c) would miss a clustered row whose only non-key
-        // data is a purgeable cell tombstone: that tombstone is still present in
-        // `surviving` at this point, so a pre-purge check would (wrongly) report
-        // surviving data and emit a phantom key-only live row after the purge.
-        let ck_names: std::collections::HashSet<&str> = clustering_key
-            .as_ref()
-            .map(|ck| ck.columns.iter().map(|(n, _)| n.as_str()).collect())
-            .unwrap_or_default();
-        let is_data_cell = |cell: &CellData| !ck_names.contains(cell.column.as_str());
-        let had_data_before = after_row_del.iter().any(is_data_cell);
-
-        // Step 3c: gc_grace / gcBefore tombstone PURGING (issue #845, parity
-        // Cassandra `8d47ebb2`). A tombstone whose on-disk `localDeletionTime`
-        // (GC-clock seconds) is STRICTLY LESS THAN `gcBefore` is purgeable and
-        // dropped from the output; one within grace (`>= gcBefore`) is retained.
-        // The boundary mirrors Cassandra exactly: `localDeletionTime < gcBefore`
-        // purges, `==` is retained.
-        //
-        // This is a SEPARATE stage that runs AFTER complex-deletion
-        // shadow-before-purge (Step 2b) and the row-tombstone / dropped-column
-        // filters above. Those stages have already removed every cell/element
-        // COVERED by a tombstone WITHIN this compaction, so purging the
-        // now-redundant marker here cannot resurrect data within this set.
-        //
-        // OVERLAP SAFETY (#921 finding 1, #935): `gc_before_secs` here is the
-        // EFFECTIVE cutoff. For a FULL compaction it is the table cutoff and
-        // `max_purgeable_timestamp` is `i64::MAX` (no non-included overlap exists),
-        // so every gc-purgeable tombstone is dropped exactly as #845. For a PARTIAL
-        // compaction WITHOUT an overlap bound `merge_partition_rows` collapsed the
-        // cutoff to `None`, so this stage is a strict no-op (every tombstone
-        // retained, #921). For a PARTIAL compaction WITH an overlap bound (#935)
-        // the cutoff is the table cutoff and `max_purgeable_timestamp` is the min
-        // write timestamp of every non-included overlapping SSTable: a tombstone is
-        // purged only when BOTH its gc grace has elapsed AND its own deletion
-        // timestamp (`markedForDeleteAt`) is STRICTLY LESS THAN that bound — so it
-        // provably shadows nothing outside the compaction set and cannot resurrect
-        // data on a later read. This mirrors Cassandra
-        // `CompactionController.getPurgeEvaluator` (`time -> time < minTimestamp`).
-        //
-        // LDT NORMALIZATION (#921 finding 2): on-disk `localDeletionTime` is an
-        // UNSIGNED 32-bit GC-clock second count (OA/DA). It is carried here as a
-        // wrapped `i32`, so a far-future LDT with bit 31 set reads as a NEGATIVE
-        // `i32`. Comparing that via `as i64` would make it look older than any
-        // `gcBefore` and purge a far-future tombstone immediately. Reinterpret
-        // the bits as unsigned (`i64::from(ldt as u32)`) before every compare.
-        if let Some(gc_before) = gc_before_secs {
-            // (a) Cell tombstones: drop any purgeable simple cell tombstone (and
-            // purgeable complex-element tombstone) from the surviving set. A cell
-            // whose `local_deletion_time` is not surfaced (`None`) is conservative-
-            // ly RETAINED — we never purge on unknown LDT (no-heuristics mandate).
-            surviving.retain(|cell| {
-                if Self::is_cell_tombstone(cell) || cell.is_deleted {
-                    // #921 finding 1: a simple cell tombstone surfaces its LDT in
-                    // its `Value::Tombstone` payload, not `CellData.local_deletion_time`
-                    // (which the reader fills only for expiring cells). Consult both
-                    // via `cell_effective_ldt` so a purgeable cell tombstone is
-                    // actually purged here — matching the survivor pre-pass.
-                    let gc_purgeable = match Self::cell_effective_ldt(cell) {
-                        Some(ldt) => i64::from(ldt as u32) < gc_before,
-                        // Unknown LDT: never purge (no-heuristics mandate).
-                        None => false,
-                    };
-                    // #935 overlap gate: the cell tombstone's own write timestamp
-                    // (`markedForDeleteAt`) must be STRICTLY BELOW the min outside
-                    // timestamp to prove it shadows nothing in a non-included
-                    // overlapping SSTable. `i64::MAX` for a full compaction lets
-                    // every gc-purgeable tombstone through unchanged.
-                    let overlap_purgeable = cell.timestamp < max_purgeable_timestamp;
-                    // RETAIN unless BOTH gates allow the purge.
-                    let keep = !(gc_purgeable && overlap_purgeable);
-                    if !keep {
-                        // True gc/overlap-safe purge of a cell tombstone (simple
-                        // or complex-element), issue #1037.
-                        purges.cell_tombstones += 1;
-                    }
-                    keep
-                } else {
-                    true
-                }
-            });
-
-            // (b) Row tombstone: a purgeable row deletion is dropped so no
-            // tombstone entry is emitted for it. Row-tombstone shadowing already
-            // ran above (`after_row_del`), so cells it covered are gone.
-            //
-            // UNKNOWN-LDT RETENTION (#921 finding 2): `row_del_ldt == 0` is the
-            // "LDT not surfaced" placeholder used by legacy/pre-V5 paths (the
-            // field is initialized to 0 and only overwritten when a real
-            // `local_deletion_time` is carried). Treat 0 as UNKNOWN and RETAIN
-            // the row tombstone — never purge on unknown LDT, matching the
-            // conservative rule already used for cell tombstones (case (a),
-            // `None => true`). Only a real, non-zero, surfaced LDT strictly
-            // below `gcBefore` purges.
-            //
-            // #935 overlap gate: also require the row deletion's own timestamp
-            // (`markedForDeleteAt` = `row_del`) to be STRICTLY BELOW the min
-            // outside timestamp so it shadows nothing in a non-included
-            // overlapping SSTable. `i64::MAX` (full compaction) is a no-op.
-            if row_del_ldt != 0
-                && i64::from(row_del_ldt as u32) < gc_before
-                && row_del.is_some_and(|d| d < max_purgeable_timestamp)
-            {
-                row_del = None;
-                // True gc/overlap-safe purge of a row tombstone (issue #1037).
-                purges.row_tombstones += 1;
-            }
-
-            // (c) Complex-deletion markers: drop each purgeable marker. The
-            // strict-supersede reduction + shadow-before-purge already ran in
-            // Step 2b, so a covered element is gone before its marker is purged.
-            // #935 overlap gate: also require the marker's `marked_for_delete_at`
-            // to be STRICTLY BELOW the min outside timestamp.
-            complex_deletions.retain(|cd| {
-                let gc_purgeable = i64::from(cd.local_deletion_time as u32) < gc_before;
-                let overlap_purgeable = cd.marked_for_delete_at < max_purgeable_timestamp;
-                let keep = !(gc_purgeable && overlap_purgeable);
-                if !keep {
-                    // True gc/overlap-safe purge of a complex-deletion marker
-                    // (issue #1037).
-                    purges.complex_deletions += 1;
-                }
-                keep
-            });
-        }
-
-        // Phantom key-only-row determination (#921 finding 3): recompute
-        // `purged_to_empty` AFTER the gc-grace cell-tombstone purge above, so a
-        // CLUSTERED row whose only non-key data was a purgeable cell tombstone —
-        // now removed from `surviving` — is recognized as data-less and emits
-        // NOTHING instead of a phantom key-only live row. `had_data_before` was
-        // captured pre-purge above; if the row HAD non-key data before the purges
-        // and has none after, it has been `purged_to_empty`. A row that was always
-        // key-only (a genuine row marker) keeps `had_data_before == false` and is
-        // preserved by Step 4 below.
-        let has_data_after = surviving.iter().any(is_data_cell);
-        let purged_to_empty = had_data_before && !has_data_after;
-
-        // Step 4: build the merged result.
-        //
-        // Attach the carried deletion metadata to whichever entry is emitted so it
-        // is not dropped by reconciliation (#886 plumbing preservation). This is
-        // behavior-neutral: the writer does not yet consume these fields.
-        let has_carried_metadata = !complex_deletions.is_empty() || range_deletion.is_some();
-
-        // Emit a live row only when real data survives. `!purged_to_empty`
-        // suppresses a clustered row whose only data was a dropped column (phantom
-        // key-only row, see above); a genuine row marker (always key-only) has
-        // `had_data_before == false` so it is preserved.
-        let built = if !surviving.is_empty() && !purged_to_empty {
-            // `surviving` is non-empty, so `max()` is `Some`; `unwrap_or(0)` only
-            // guards the type and never triggers.
-            let row_ts = surviving.iter().map(|c| c.timestamp).max().unwrap_or(0);
-            let live = MergeEntry::new(
-                run_index,
-                key,
-                clustering_key,
-                row_ts,
-                RowData::Live { cells: surviving },
-            );
-            // Issue #932: a row deletion that survived purging COEXISTS with the
-            // surviving (strictly-newer) cells. Carry it on the live entry so the
-            // merge→mutation step emits a `HAS_DELETION` row holding both — the row
-            // deletion keeps shadowing older cells of OTHER columns living in
-            // SSTables not part of a partial compaction, preventing resurrection.
-            // Step 3 already dropped the cells this deletion covers (ts <= row_del),
-            // so the deletion shadows nothing within `surviving`.
-            Some(match row_del {
-                Some(deletion_time) => live.with_row_deletion(deletion_time, row_del_ldt),
-                None => live,
-            })
-        } else if let Some(deletion_time) = row_del {
-            // No surviving data. If a row tombstone exists, keep the row shadowed
-            // so downstream still emits the deletion (preserves #505/#498 absence).
-            Some(MergeEntry::new(
-                run_index,
-                key,
-                clustering_key,
-                deletion_time,
-                RowData::Tombstone {
-                    deletion_time,
-                    // Preserve the source LDT paired with the winning deletion_time
-                    // (#873) instead of resetting it to 0; the writer encodes it
-                    // verbatim and gc_grace decisions stay faithful.
-                    local_deletion_time: row_del_ldt,
-                },
-            ))
-        } else if has_carried_metadata {
-            // No surviving data AND no row tombstone, but the cluster still carries
-            // complex/range deletion metadata. Emit a metadata-only entry (an empty
-            // `Live` row) so it survives reconciliation and reaches downstream
-            // consumers (#844/#846/#899).
-            Some(MergeEntry::new(
-                run_index,
-                key,
-                clustering_key,
-                0,
-                RowData::Live { cells: vec![] },
-            ))
-        } else {
-            // Truly empty/absent row.
-            None
-        };
-
-        built.map(|entry| {
-            let entry = if complex_deletions.is_empty() {
-                entry
-            } else {
-                entry.with_complex_deletions(complex_deletions)
-            };
-            match range_deletion {
-                Some(rd) => entry.with_range_deletion(rd),
-                None => entry,
-            }
-        })
+        // Step 2b: complex-deletion strict-supersede + shadow-before-purge.
+        state.apply_complex_deletions();
+        // Step 3: row-tombstone shadowing.
+        state.shadow_by_row_deletion();
+        // Step 3b: dropped-column filtering (captures the phantom-row guard).
+        state.filter_dropped_columns(dropped_columns);
+        // Step 3c: gc_grace / overlap-aware tombstone purging (tallies #1037).
+        state.purge_gc_grace(gc_before_secs, max_purgeable_timestamp, purges);
+        // Step 4: phantom-row guard + emit the merged entry.
+        state.build()
     }
 
     /// Convert reconciled `CellData`s into writer `CellOperation`s (epic #899,
